requirements.txt
pandas
requests
openmeteo-requests
requests-cache
supabase
python-dotenv